app/requirements.txt
Flask==2.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.3